app/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
groq
pydantic
spacy>=3.7.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-multipart>=0.0.6